sudan-news-pipeline/requirements.txt
# Core dependencies
click>=8.0.0
python-dotenv>=0.19.0

# Heavy ML libraries
torch>=2.0.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
transformers>=4.21.0
huggingface-hub>=0.15.0

# System monitoring
psutil>=5.9.0

# Data processing
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
feedparser>=6.0.0
